Create name of aggregate expression from expressions
