+++ datafusion/physical-plan/src/aggregates/mod.rs
@@ -2025,7 +2025,6 @@ mod tests {
             &sort_exprs,
             &ordering_req,
             dfschema,
-            "LAST_VALUE(b)",
             false,
             false,
             false,
@@ -2258,7 +2257,6 @@ mod tests {
                     &sort_exprs,
                     &ordering_req,
                     &test_df_schema,
-                    "array_agg",
                     false,
                     false,
                     false,
