+++ datafusion/physical-plan/src/windows/mod.rs
@@ -146,7 +146,6 @@ pub fn create_window_expr(
 
             let aggregate = AggregateExprBuilder::new(Arc::clone(fun), args.to_vec())
                 .schema(Arc::new(input_schema.clone()))
-                .name(name)
                 .order_by(order_by.to_vec())
                 .sort_exprs(sort_exprs)
                 .with_ignore_nulls(ignore_nulls)
